Separate commands for synaptor
